Adding AzureAD Preview generation

+++ build/Create-CompatModule.ps1
@@ -7,7 +7,10 @@ param($TargetDirectory = $null, $Module = "AzureAD", [switch] $noclean)
 . (join-path $psscriptroot "/common-functions.ps1")
 . (join-path $psscriptroot "../src/CompatibilityAdapter.ps1")
 
-Remove-BuildDirectories
+if($noclean -eq $true) {
+    Remove-BuildDirectories
+}
+
 $mapper = [CompatibilityAdapterBuilder]::new($Module)
 
 $customizationFiles = Get-CustomizationFiles -Module $Module
@@ -15,7 +18,6 @@ foreach($file in $customizationFiles){
     $cmds = & $file
     $mapper.AddCustomization($cmds)
 }
-
 $AdditionalFunctions = Get-CustomizationFiles -Directory 'AdditionalFunctions'  -Module $Module
 foreach($file in $AdditionalFunctions){
     $mapper.AddHelperCommand($file)
